backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
PyJWT
firebase-admin
